Fixed code coverage to work on recent Xcode
My machine updated to the most recent copy of Xcode and the code
coverage rules in CppUTest stopped working.  It appears that the llvm
version of gcov doesn't want to take multiple source filenames on the
command line so I added a for loop to iterate through each one
manually.